backend/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["V-U-Simon <V-U-Simon@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
djangorestframework-camel-case = "^1.4.2"
django-filter = "^23.2"
pytest = "^7.4.1"
pytest-django = "^4.5.2"
drf-yasg = "^1.21.7"
graphene-django = "^3.1.5"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
